+++ zkevm-circuits/src/mpt_circuit/branch.rs
@@ -18,7 +18,7 @@ use crate::{
     mpt_circuit::{
         helpers::{nibble_rlc, Indexable, KECCAK},
         param::{HASH_WIDTH, RLP_NIL},
-        MPTConfig, MPTState, RlpItemType,
+        MPTConfig, MptMemory, RlpItemType,
     },
 };
 
@@ -273,7 +273,7 @@ impl<F: Field> BranchGadget<F> {
         &self,
         region: &mut CachedRegion<'_, '_, F>,
         _mpt_config: &MPTConfig<F>,
-        _pv: &mut MPTState<F>,
+        _memory: &mut MptMemory<F>,
         offset: usize,
         is_placeholder: &[bool; 2],
         key_rlc: &mut F,
