+++ zkevm-circuits/src/mpt_circuit/extension_branch.rs
@@ -16,7 +16,7 @@ use crate::{
     mpt_circuit::{
         helpers::{key_memory, parent_memory, Indexable, KeyData, ParentData},
         witness_row::ExtensionBranchRowType,
-        MPTConfig, MPTState,
+        MPTConfig, MptMemory,
     },
 };
 
@@ -59,7 +59,6 @@ impl<F: Field> ExtensionBranchConfig<F> {
             // Load the parent values
             for is_s in [true, false] {
                 config.parent_data[is_s.idx()] = ParentData::load(
-                    "branch load",
                     cb,
                     &mut ctx.memory[parent_memory(is_s)],
                     0.expr(),
@@ -184,7 +183,7 @@ impl<F: Field> ExtensionBranchConfig<F> {
         &self,
         region: &mut CachedRegion<'_, '_, F>,
         mpt_config: &MPTConfig<F>,
-        pv: &mut MPTState<F>,
+        memory: &mut MptMemory<F>,
         offset: usize,
         node: &Node,
         rlp_values: &[RLPItemWitness],
@@ -196,13 +195,13 @@ impl<F: Field> ExtensionBranchConfig<F> {
 
         let key_data =
             self.key_data
-                .witness_load(region, offset, &mut pv.memory[key_memory(true)], 0)?;
+                .witness_load(region, offset, &mut memory[key_memory(true)], 0)?;
         let mut parent_data = vec![ParentDataWitness::default(); 2];
         for is_s in [true, false] {
             parent_data[is_s.idx()] = self.parent_data[is_s.idx()].witness_load(
                 region,
                 offset,
-                &mut pv.memory[parent_memory(is_s)],
+                &mut memory[parent_memory(is_s)],
                 0,
             )?;
             self.is_placeholder[is_s.idx()].assign(
@@ -222,7 +221,7 @@ impl<F: Field> ExtensionBranchConfig<F> {
             self.extension.assign(
                 region,
                 mpt_config,
-                pv,
+                memory,
                 offset,
                 &key_data,
                 &mut key_rlc,
@@ -239,7 +238,7 @@ impl<F: Field> ExtensionBranchConfig<F> {
             self.branch.assign(
                 region,
                 mpt_config,
-                pv,
+                memory,
                 offset,
                 &extension_branch.is_placeholder,
                 &mut key_rlc,
@@ -256,7 +255,7 @@ impl<F: Field> ExtensionBranchConfig<F> {
                 KeyData::witness_store(
                     region,
                     offset,
-                    &mut pv.memory[key_memory(is_s)],
+                    &mut memory[key_memory(is_s)],
                     key_rlc_post_branch,
                     key_mult_post_branch,
                     num_nibbles,
@@ -267,7 +266,7 @@ impl<F: Field> ExtensionBranchConfig<F> {
                 ParentData::witness_store(
                     region,
                     offset,
-                    &mut pv.memory[parent_memory(is_s)],
+                    &mut memory[parent_memory(is_s)],
                     mod_node_hash_rlc[is_s.idx()],
                     false,
                     false,
@@ -277,7 +276,7 @@ impl<F: Field> ExtensionBranchConfig<F> {
                 KeyData::witness_store(
                     region,
                     offset,
-                    &mut pv.memory[key_memory(is_s)],
+                    &mut memory[key_memory(is_s)],
                     key_data.rlc,
                     key_data.mult,
                     key_data.num_nibbles,
@@ -288,7 +287,7 @@ impl<F: Field> ExtensionBranchConfig<F> {
                 ParentData::witness_store(
                     region,
                     offset,
-                    &mut pv.memory[parent_memory(is_s)],
+                    &mut memory[parent_memory(is_s)],
                     parent_data[is_s.idx()].rlc,
                     parent_data[is_s.idx()].is_root,
                     true,
